backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openai
python-dotenv
python-dateutil
psycopg2-binary
